requirements.txt
gradio
pymupdf
ultralytics
pillow
reportlab
